Add base page object
This will be used in feature tests to help make interacting with the
page easier. Page objects contain all the logic for an individual page
(or section if it needs to be broken out). This allows for more modular
testing.

See a blog post by my one of my former coworkers for more insight on why
we typically use this pattern: https://teamgaslight.com/blog/6-ways-to-remove-pain-from-feature-testing-in-ruby-on-rails